C. Prepend and Append
Timur initially had a binary string† s (possibly of length 0). He performed the following operation several (possibly zero) times:
Add 0 to one end of the string and 1 to the other end of the string. For example, starting from the string 1011, you can obtain either 010111 or 110110.
You are given Timur's final string. What is the length of the shortest possible string he could have started with?
† A binary string is a string (possibly the empty string) whose characters are either 0 or 1.